move placeholder to data directory
